feat: reset() cannot be used safely; remove it
BREAKING CHANGE: reset() is gone. Just don't use it.

It is not really safe to use reset() ever. This library is
shared between different bits of code. You cannot know if
they might need to yield, even if you didn't.

It's unclear it's even useful in other cases; much of the cost
of using the library is in checking, not actually spinning.